+++ python_modules/libraries/dagster-dg/dagster_dg/cli/__init__.py
@@ -2,12 +2,14 @@ from pathlib import Path
 
 import click
 
-from dagster_dg.cli.code_location import code_location_group
-from dagster_dg.cli.component import component_group
-from dagster_dg.cli.component_type import component_type_group
-from dagster_dg.cli.deployment import deployment_group
+from dagster_dg.cli.check import check_group
+from dagster_dg.cli.configure_editor import configure_editor_group
 from dagster_dg.cli.dev import dev_command
+from dagster_dg.cli.docs import docs_group
 from dagster_dg.cli.global_options import dg_global_options
+from dagster_dg.cli.info import info_group
+from dagster_dg.cli.list import list_group
+from dagster_dg.cli.scaffold import scaffold_group
 from dagster_dg.component import RemoteComponentRegistry
 from dagster_dg.config import normalize_cli_config
 from dagster_dg.context import DgContext
@@ -21,10 +23,12 @@ def create_dg_cli():
     @click.group(
         name="dg",
         commands={
-            "code-location": code_location_group,
-            "component": component_group,
-            "component-type": component_type_group,
-            "deployment": deployment_group,
+            "check": check_group,
+            "docs": docs_group,
+            "info": info_group,
+            "configure-editor": configure_editor_group,
+            "list": list_group,
+            "scaffold": scaffold_group,
             "dev": dev_command,
         },
         context_settings={
