+++ python_modules/libraries/dagster-dg/dagster_dg/utils/__init__.py
@@ -341,7 +341,7 @@ def _format_error_message(message: str) -> str:
 
 def generate_missing_component_type_error_message(component_key_str: str) -> str:
     return f"""
-        No component type `{component_key_str}` is registered. Use 'dg component-type list'
+        No component type `{component_key_str}` is registered. Use 'dg list component-type'
         to see the registered component types in your environment. You may need to install a package
         that provides `{component_key_str}` into your environment.
     """
