+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/component_command_tests/test_check.py
@@ -93,7 +93,7 @@ def test_component_check_succeeds_non_default_component_package() -> None:
         assert BASIC_VALID_VALUE.component_type_filepath
         shutil.copy(BASIC_VALID_VALUE.component_type_filepath, components_dir / "__init__.py")
 
-        result = runner.invoke("component", "check")
+        result = runner.invoke("check", "component")
         assert_runner_result(result, exit_0=True)
 
 
@@ -115,7 +115,7 @@ def test_validation_cli(test_case: ComponentValidationTestCase) -> None:
         ) as tmpdir,
     ):
         with pushd(tmpdir):
-            result = runner.invoke("component", "check")
+            result = runner.invoke("check", "component")
             if test_case.should_error:
                 assert result.exit_code != 0, str(result.stdout)
 
@@ -149,8 +149,8 @@ def test_validation_cli_multiple_components(scope_check_run: bool) -> None:
     ):
         with pushd(str(tmpdir)):
             result = runner.invoke(
-                "component",
                 "check",
+                "component",
                 *(
                     [
                         str(Path("foo_bar") / "components" / "basic_component_missing_value"),
@@ -180,8 +180,8 @@ def test_validation_cli_multiple_components_filter() -> None:
     ):
         with pushd(tmpdir):
             result = runner.invoke(
-                "component",
                 "check",
+                "component",
                 str(Path("foo_bar") / "components" / "basic_component_missing_value"),
             )
             assert result.exit_code != 0, str(result.stdout)
@@ -206,7 +206,7 @@ def test_validation_cli_local_component_cache() -> None:
         ) as code_location_dir,
     ):
         with pushd(code_location_dir):
-            result = runner.invoke("component", "check")
+            result = runner.invoke("check", "component")
             assert re.search(
                 r"CACHE \[write\].*basic_component_success.*local_component_registry", result.stdout
             )
@@ -216,7 +216,7 @@ def test_validation_cli_local_component_cache() -> None:
             )
 
             # Local components should all be cached
-            result = runner.invoke("component", "check")
+            result = runner.invoke("check", "component")
             assert not re.search(
                 r"CACHE \[write\].*basic_component_success.*local_component_registry", result.stdout
             )
@@ -242,7 +242,7 @@ def test_validation_cli_local_component_cache() -> None:
             ).write_text(contents + "\n")
 
             # basic_component_success local component is now be invalidated and needs to be re-cached, the other one should still be cached
-            result = runner.invoke("component", "check")
+            result = runner.invoke("check", "component")
             assert re.search(
                 r"CACHE \[write\].*basic_component_success.*local_component_registry", result.stdout
             )
