+++ python_modules/libraries/dagster-dg/dagster_dg_tests/test_cache.py
@@ -16,61 +16,61 @@ cache_runner_args = {"verbose": True}
 
 def test_load_from_cache():
     with ProxyRunner.test(**cache_runner_args) as runner, example_code_location(runner):
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [hit]" in result.output
 
 
 def test_cache_invalidation_uv_lock():
     with ProxyRunner.test(**cache_runner_args) as runner, example_code_location(runner):
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
         subprocess.run(["uv", "add", "dagster-components[dbt]", "dagster-dbt"], check=True)
 
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
 
 
 def test_cache_invalidation_modified_lib():
     with ProxyRunner.test(**cache_runner_args) as runner, example_code_location(runner):
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
-        result = runner.invoke("component-type", "scaffold", "my_component")
+        result = runner.invoke("scaffold", "component-type", "my_component")
         assert_runner_result(result)
 
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
 
 
 def test_cache_no_invalidation_modified_pkg():
     with ProxyRunner.test(**cache_runner_args) as runner, example_code_location(runner):
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
         Path("foo_bar/submodule.py").write_text("print('hello')")
 
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [hit]" in result.output
 
 
 def test_clear_cache():
     with ProxyRunner.test(**cache_runner_args) as runner, example_code_location(runner):
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
@@ -78,7 +78,7 @@ def test_clear_cache():
         result = runner.invoke("--clear-cache")
         assert_runner_result(result)
         assert "CACHE [clear-all]" in result.output
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
 
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
@@ -94,7 +94,7 @@ def test_rebuild_component_registry_success():
         assert_runner_result(result)
         assert "CACHE [clear-key]" in result.output
 
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE [hit]" in result.output
 
@@ -104,7 +104,7 @@ def test_rebuild_component_registry_fails_with_subcommand():
         ProxyRunner.test(**cache_runner_args) as runner,
         isolated_example_code_location_foo_bar(runner),
     ):
-        result = runner.invoke("--rebuild-component-registry", "component-type", "list")
+        result = runner.invoke("--rebuild-component-registry", "list", "component-type")
         assert_runner_result(result, exit_0=False)
         assert "Cannot specify --rebuild-component-registry with a subcommand." in result.output
 
@@ -128,6 +128,6 @@ def test_cache_disabled():
         ProxyRunner.test(**cache_runner_args, disable_cache=True) as runner,
         example_code_location(runner),
     ):
-        result = runner.invoke("component-type", "list")
+        result = runner.invoke("list", "component-type")
         assert_runner_result(result)
         assert "CACHE" not in result.output
